+++ Model/ComponentInterface2/CMPData.cpp
@@ -5,11 +5,11 @@ void EXPORT getKindAndArray(const std::string& ddml,
                      std::string& kind, bool& isArray)
    {
    isArray = (ddml.find("array=\"T\"") != string::npos);
-   unsigned posKind = ddml.find("kind=\"");
+   size_t posKind = ddml.find("kind=\"");
    if (posKind != string::npos)
       {
       kind = ddml.substr(posKind + strlen("kind=\""));
-      unsigned posQuote = kind.find('\"');
+      size_t posQuote = kind.find('\"');
       if (posQuote == string::npos)
          throw runtime_error("Invalid data type string: " + ddml);
       kind.erase(posQuote);
